refactor: Move 'Profile' to be on its own folder
