+++ web/components/Pages/Profile/Infobox.tsx
@@ -1,10 +1,4 @@
 import {
-  AlertDialog,
-  AlertDialogBody,
-  AlertDialogContent,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogOverlay,
   Button,
   Grid,
   Heading,
@@ -14,15 +8,16 @@ import {
   VStack,
 } from '@chakra-ui/react';
 import { useRouter } from 'next/router';
-import { memo, useRef, useState } from 'react';
-import { FaEraser, FaFire, FaPencilRuler } from 'react-icons/fa';
+import { memo, useState } from 'react';
+import { FaEraser, FaPencilRuler } from 'react-icons/fa';
 
-import { useStatusAndUser } from '../../utils/hooks';
-import axios from '../../utils/http';
-import routes from '../../utils/routes';
-import type { Status, User } from '../../utils/types';
-import TextInput from '../Input/TextInput';
-import { FailedToast, SuccessToast } from '../Toast';
+import { useStatusAndUser } from '../../../utils/hooks';
+import axios from '../../../utils/http';
+import routes from '../../../utils/routes';
+import type { Status, User } from '../../../utils/types';
+import TextInput from '../../Input/TextInput';
+import AlertOverlay from '../../Overlay/AlertOverlay';
+import { FailedToast, SuccessToast } from '../../Toast';
 
 /**
  * Edit profile box to edit non-sensitive information.
@@ -37,7 +32,6 @@ const Infobox = ({ status, user }: { status: Status; user: User }) => {
   const [isLoading, setIsLoading] = useState(false);
   const [isOpen, setIsOpen] = useState(false);
   const [phoneNumber, setPhoneNumber] = useState(user.phoneNumber);
-  const cancelRef = useRef(null);
   const router = useRouter();
   const toast = useToast();
 
@@ -86,40 +80,13 @@ const Infobox = ({ status, user }: { status: Status; user: User }) => {
 
   return (
     <>
-      <AlertDialog
+      <AlertOverlay
+        title="Delete Profile"
+        description="Are you sure? This is a potentially destructive and irreversible action."
         isOpen={isOpen}
-        leastDestructiveRef={cancelRef}
-        onClose={() => setIsOpen(false)}
-        isCentered
-      >
-        <AlertDialogOverlay>
-          <AlertDialogContent>
-            <AlertDialogHeader fontSize="lg" fontWeight="bold">
-              Delete User
-            </AlertDialogHeader>
-
-            <AlertDialogBody>
-              Are you sure? This is a potentially destructive and irreversible
-              action.
-            </AlertDialogBody>
-
-            <AlertDialogFooter>
-              <Button ref={cancelRef} onClick={() => setIsOpen(false)}>
-                Cancel
-              </Button>
-
-              <Button
-                colorScheme="red"
-                leftIcon={<FaFire />}
-                onClick={deleteProfile}
-                ml={3}
-              >
-                Delete
-              </Button>
-            </AlertDialogFooter>
-          </AlertDialogContent>
-        </AlertDialogOverlay>
-      </AlertDialog>
+        setIsOpen={setIsOpen}
+        fn={deleteProfile}
+      />
 
       <VStack as="section" p={[2, 10]} spacing={5} mt={10}>
         <Heading as="p" size="lg">
@@ -134,6 +101,7 @@ const Infobox = ({ status, user }: { status: Status; user: User }) => {
           w="full"
           templateColumns={{ lg: 'repeat(2, 1fr)' }}
           gap={[5, 10]}
+          onSubmit={editProfile}
         >
           <TextInput
             label="Full name"
