+++ web/pages/profile.tsx
@@ -2,12 +2,12 @@ import { memo } from 'react';
 
 import AuthRoute from '../components/AuthRoute';
 import Layout from '../components/Layout';
-import Attendancebox from '../components/Profile/Attendancebox';
-import Authenticatorbox from '../components/Profile/Authenticatorbox';
-import Infobox from '../components/Profile/Infobox';
-import PasswordBox from '../components/Profile/Passwordbox';
-import Sessionbox from '../components/Profile/Sessionbox';
-import Topbox from '../components/Profile/Topbox';
+import Attendancebox from '../components/Pages/Profile/Attendancebox';
+import Authenticatorbox from '../components/Pages/Profile/Authenticatorbox';
+import Infobox from '../components/Pages/Profile/Infobox';
+import PasswordBox from '../components/Pages/Profile/Passwordbox';
+import Sessionbox from '../components/Pages/Profile/Sessionbox';
+import Topbox from '../components/Pages/Profile/Topbox';
 import { useMe } from '../utils/hooks';
 
 /**
@@ -27,11 +27,11 @@ const Profile = () => {
             <Infobox status={data.status} user={data.status.user} />
             <PasswordBox />
             <Authenticatorbox user={data.status.user} status={data.status} />
+
+            {data.attendance && <Attendancebox attendances={data.attendance} />}
+            {data.sessions && <Sessionbox sessions={data.sessions} />}
           </>
         )}
-
-        {data.attendance && <Attendancebox attendances={data.attendance} />}
-        {data.sessions && <Sessionbox sessions={data.sessions} />}
       </Layout>
     </AuthRoute>
   );
